requirements.txt
alembic==1.13.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Limiter==3.5.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.10.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pytest==8.1.1
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
rich==13.7.1
soupsieve==2.5
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
